No. of subsequence of '123' in the first 100,000 digits of pi
Approach:
Build a tree. Consider results[0] as root node. Iterate along the string
if the character is '1', add a node to the the root of the tree. If the
character is '2', add a child to each first level node. If the character
is '3', add a child to each second level node. Return the number of
third layer nodes.